feat: added client example
