+++ packages/fuel-streams/package.json
@@ -13,7 +13,7 @@
   },
   "files": ["dist"],
   "scripts": {
-    "test-consume": "tsx ./examples/consume/index.ts",
+    "test-client": "tsx ./examples/client/index.ts",
     "build": "tsup",
     "build:watch": "tsup --watch",
     "ts:check": "tsc --noEmit"
